+++ include/build_parameters.hpp
@@ -1,6 +1,6 @@
 #ifndef BUILD_PARAMETERS_HPP
 #define BUILD_PARAMETERS_HPP
-namespace buildparams {
+namespace raytracer::buildparams {
 
 #ifdef USE_SINGLE_TASK
 constexpr bool use_single_task = true;
