+++ include/constant_medium.hpp
@@ -5,7 +5,8 @@
 #include "material.hpp"
 #include "sphere.hpp"
 #include "texture.hpp"
-#include "visit.hpp"
+
+namespace raytracer::scene {
 
 using hittableVolume_t = std::variant<std::monostate, sphere, box>;
 
@@ -25,54 +26,9 @@ class constant_medium {
       , neg_inv_density { -1 / d }
       , phase_function { isotropic_material { a } } {}
 
-  bool hit(auto& ctx, const ray& r, real_t min, real_t max, hit_record& rec,
-           material_t& hit_material_type) const {
-    auto& rng = ctx.rng;
-    hit_material_type = phase_function;
-    material_t temp_material_type;
-    hit_record rec1, rec2;
-    if (!dev_visit(hittable_hit_visitor(ctx, r, -infinity, infinity, rec1,
-                                        temp_material_type),
-                   boundary)) {
-      return false;
-    }
-
-    if (!dev_visit(hittable_hit_visitor(ctx, r, rec1.t + 0.0001f, infinity, rec2,
-                             temp_material_type),
-            boundary)) {
-      return false;
-    }
-
-    if (rec1.t < min)
-      rec1.t = min;
-    if (rec2.t > max)
-      rec2.t = max;
-    if (rec1.t >= rec2.t)
-      return false;
-    if (rec1.t < 0)
-      rec1.t = 0;
-
-    const auto ray_length = sycl::length(r.direction());
-    /// Distance between the two hitpoints affect of probability
-    /// of the ray hitting a smoke particle
-    const auto distance_inside_boundary = (rec2.t - rec1.t) * ray_length;
-    const auto hit_distance = neg_inv_density * sycl::log(rng.real());
-
-    /// With lower density, hit_distance has higher probabilty
-    /// of being greater than distance_inside_boundary
-    if (hit_distance > distance_inside_boundary)
-      return false;
-
-    rec.t = rec1.t + hit_distance / ray_length;
-    rec.p = r.at(rec.t);
-
-    rec.normal = vec { 1, 0, 0 }; // arbitrary
-    rec.front_face = true;        // also arbitrary
-    return true;
-  }
-
   hittableVolume_t boundary;
   real_t neg_inv_density;
   material_t phase_function;
 };
+} // namespace raytracer::scene
 #endif
