+++ include/material.hpp
@@ -1,31 +1,18 @@
 #ifndef RT_SYCL_MATERIAL_HPP
 #define RT_SYCL_MATERIAL_HPP
 
-#include <iostream>
+#include <variant>
 
-#include "hit_record.hpp"
+#include "primitives.hpp"
 #include "texture.hpp"
-#include "vec.hpp"
-#include "visit.hpp"
 
+namespace raytracer::scene {
 struct lambertian_material {
   lambertian_material() = default;
   lambertian_material(const color& a)
       : albedo { solid_texture { a } } {}
   lambertian_material(const texture_t& a)
       : albedo { a } {}
-
-  bool scatter(auto& ctx, const ray& r_in, const hit_record& rec,
-               color& attenuation, ray& scattered) const {
-    auto& rng = ctx.rng;
-    vec scatter_direction = rec.normal + rng.unit_vec();
-    scattered = ray(rec.p, scatter_direction, r_in.time());
-    // Attenuation of the ray hitting the object is modified based on the color
-    // at hit point
-    attenuation *= dev_visit(texture_value_visitor(ctx, rec), albedo);
-    return true;
-  }
-  color emitted(auto&, const hit_record& rec) { return color(0, 0, 0); }
   texture_t albedo;
 };
 
@@ -35,18 +22,6 @@ struct metal_material {
       : albedo { a }
       , fuzz { std::clamp(f, 0.0f, 1.0f) } {}
 
-  bool scatter(auto& ctx, const ray& r_in, const hit_record& rec,
-               color& attenuation, ray& scattered) const {
-    auto& rng = ctx.rng;
-    vec reflected = reflect(unit_vector(r_in.direction()), rec.normal);
-    scattered = ray(rec.p, reflected + fuzz * rng.in_unit_ball(), r_in.time());
-    // Attenuation of the ray hitting the object is modified based on the color
-    // at hit point
-    attenuation *= albedo;
-    return (dot(scattered.direction(), rec.normal) > 0);
-  }
-
-  color emitted(auto&, const hit_record& rec) { return color(0, 0, 0); }
   color albedo;
   real_t fuzz;
 };
@@ -64,28 +39,6 @@ struct dielectric_material {
     return r0 + (1 - r0) * sycl::pow((1 - cosine), 5.0f);
   }
 
-  bool scatter(auto& ctx, const ray& r_in, const hit_record& rec,
-               color& attenuation, ray& scattered) const {
-    // Attenuation of the ray hitting the object is modified based on the color
-    // at hit point
-    auto& rng = ctx.rng;
-    attenuation *= albedo;
-    real_t refraction_ratio = rec.front_face ? (1.0f / ref_idx) : ref_idx;
-    vec unit_direction = unit_vector(r_in.direction());
-    real_t cos_theta = sycl::fmin(-sycl::dot(unit_direction, rec.normal), 1.0f);
-    real_t sin_theta = sycl::sqrt(1.0f - cos_theta * cos_theta);
-    bool cannot_refract = refraction_ratio * sin_theta > 1.0f;
-    vec direction;
-    if (cannot_refract || reflectance(cos_theta, refraction_ratio) > rng.real())
-      direction = reflect(unit_direction, rec.normal);
-    else
-      direction = refract(unit_direction, rec.normal, refraction_ratio);
-
-    scattered = ray(rec.p, direction, r_in.time());
-    return true;
-  }
-
-  color emitted(auto&, const hit_record& rec) { return color(0, 0, 0); }
   // Refractive index of the glass
   real_t ref_idx;
   // Color of the glass
@@ -99,12 +52,6 @@ struct lightsource_material {
   lightsource_material(const color& a)
       : emit { solid_texture { a } } {}
 
-  template <typename... T> bool scatter(T&...) const { return false; }
-
-  color emitted(auto& ctx, const hit_record& rec) {
-    return dev_visit(texture_value_visitor(ctx, rec), emit);
-  }
-
   texture_t emit;
 };
 
@@ -113,69 +60,12 @@ struct isotropic_material {
       : albedo { solid_texture { a } } {}
   isotropic_material(texture_t& a)
       : albedo { a } {}
-
-  bool scatter(auto& ctx, const ray& r_in, const hit_record& rec,
-               color& attenuation, ray& scattered) const {
-    auto& rng = ctx.rng;
-    scattered = ray(rec.p, rng.in_unit_ball(), r_in.time());
-    attenuation *= dev_visit(texture_value_visitor(ctx, rec), albedo);
-    return true;
-  }
-
-  color emitted(auto&, const hit_record& rec) { return color(0, 0, 0); }
-
   texture_t albedo;
 };
 
 using material_t =
     std::variant<std::monostate, lambertian_material, metal_material,
                  dielectric_material, lightsource_material, isotropic_material>;
-
-struct material_emitted_visitor {
- private:
-  task_context& ctx;
-  const hit_record& rec;
-
- public:
-  material_emitted_visitor(task_context& ctx, hit_record& rec)
-      : ctx { ctx }
-      , rec { rec } {}
-
-  template <typename M> color operator()(M&& material) {
-    return material.emitted(ctx, rec);
-  }
-
-  color operator()(std::monostate) {
-    assert(false && "unreachable");
-    return { 0.f, 0.f, 0.f };
-  }
-};
-
-struct material_scatter_visitor {
- private:
-  task_context& ctx;
-  const ray& r_in;
-  const hit_record& rec;
-  color& attenuation;
-  ray& scattered;
-
- public:
-  material_scatter_visitor(auto& ctx, const ray& r_in, const hit_record& rec,
-                           color& attenuation, ray& scattered)
-      : ctx { ctx }
-      , r_in { r_in }
-      , rec { rec }
-      , attenuation { attenuation }
-      , scattered { scattered } {}
-
-  template <typename M> bool operator()(M&& material) {
-    return material.scatter(ctx, r_in, rec, attenuation, scattered);
-  }
-
-  bool operator()(std::monostate) {
-    assert(false && "unreachable");
-    return false;
-  }
-};
+} // namespace raytracer::scene
 
 #endif
