+++ include/texture.hpp
@@ -1,54 +1,42 @@
 #ifndef RT_SYCL_TEXTURE_HPP
 #define RT_SYCL_TEXTURE_HPP
-#include "hit_record.hpp"
-#include "rtweekend.hpp"
-#include "vec.hpp"
-#include <array>
-#include <cmath>
-#include <iostream>
-#include <optional>
+#include <tuple>
+#include <variant>
 #include <vector>
-
-#include "sycl.hpp"
 
 #define STB_IMAGE_IMPLEMENTATION
 #include <stb/stb_image.h>
 
+#include "primitives.hpp"
+#include "sycl.hpp"
+
+namespace raytracer::scene {
 // Solid texture consists of a single color
 struct solid_texture {
+ private:
+  color color_value;
+
+ public:
   solid_texture() = default;
   solid_texture(const color& c)
       : color_value { c } {}
   solid_texture(real_t red, real_t green, real_t blue)
       : solid_texture { color { red, green, blue } } {}
-  // For solid texture, the color is same throughout the sphere
-  color value(auto&, const hit_record&) const { return color_value; }
-
- private:
-  color color_value;
+  color get_color() const { return color_value; }
 };
 
 // Takes two solid_textures to create checker pattern
 struct checker_texture {
-  checker_texture() = default;
+ private:
+  color odd;
+  color even;
 
-  checker_texture(const solid_texture& x, const solid_texture& y)
-      : odd { x }
-      , even { y } {}
+ public:
+  checker_texture() = default;
   checker_texture(const color& c1, const color& c2)
-      : odd { solid_texture { c1 } }
-      , even { solid_texture { c2 } } {}
-  // Color value is different based on normalised spherical coordinates
-  color value(auto& ctx, const hit_record& rec) const {
-    auto sines = sycl::sin(10 * rec.p.x()) * sycl::sin(10 * rec.p.y()) *
-                 sycl::sin(10 * rec.p.z());
-    if (sines < 0)
-      return odd.value(ctx, rec);
-    else
-      return even.value(ctx, rec);
-  }
-  solid_texture odd;
-  solid_texture even;
+      : odd { c1 }
+      , even { c2 } {}
+  auto get_colors() const { return std::tuple(odd, even); };
 };
 
 /**
@@ -130,48 +118,37 @@ struct image_texture {
                                       { texture_data.size() / 3, 3 } };
   }
 
-  /// Get the color for the texture at the given place
-  /// \todo rename this value() to color() everywhere?
-  color value(auto& ctx, const hit_record& rec) const {
+  /**
+    @brief Get the color for normalised coordinate
+
+    @param u normalised X coordinate
+    @param v normalised Y coordinate
+    @param base_ptr pointer to the image texture global buffer
+    @return color
+   */
+  color color_at(real_t u, real_t v,
+                 sycl::global_ptr<uint8_t> texture_data) const {
     // If texture data is unavailable, return solid cyan
     // The image is repeated by the repetition factor
 
-    std::size_t i =
-        sycl::fmod(rec.u * cyclic_frequency, (real_t)1) * (width - 1);
+    std::size_t i = sycl::fmod(u * cyclic_frequency, (real_t)1) * (width - 1);
     // The image frame buffer is going downwards, so flip the y axis
     std::size_t j =
-        (1 - sycl::fmod(rec.v * cyclic_frequency, (real_t)1)) * (height - 1);
+        (1 - sycl::fmod(v * cyclic_frequency, (real_t)1)) * (height - 1);
     std::size_t local_offset = j * width + i;
     std::size_t pix_idx = local_offset + offset;
     auto scale = 1.f / 255;
-    auto& texture_data = ctx.texture_data;
     return { texture_data[pix_idx * 3] * scale,
              texture_data[pix_idx * 3 + 1] * scale,
              texture_data[pix_idx * 3 + 2] * scale };
   }
 };
 
-using texture_t = std::variant<std::monostate, checker_texture, solid_texture, image_texture>;
-
-struct texture_value_visitor {
-  private:
-    task_context& ctx;
-    const hit_record& rec;
-
-  public:
-    texture_value_visitor(task_context& ctx, const hit_record& rec):ctx{ctx}, rec{rec}{}
-    template<typename T>
-    color operator()(T&& texture) {
-      return texture.value(ctx, rec);
-    }
-
-    color operator()(std::monostate) {
-      assert(false && "unreachable");
-      return {0.f,0.f,0.f};
-    }
-};
+using texture_t =
+    std::variant<std::monostate, checker_texture, solid_texture, image_texture>;
 
 // Start filled with the fallback texture (solid blue) for texture load error
 std::vector<uint8_t> image_texture::texture_data { 0, 0, 1 };
 bool image_texture::frozen = false;
+} // namespace raytracer::scene
 #endif
